Move helper class to spring package so that loadClass can find it
