+++ instrumentation/spring/spring-webmvc-3.1/javaagent/src/main/java/io/opentelemetry/javaagent/instrumentation/springwebmvc/SpringWebMvcInstrumentationModule.java
@@ -19,6 +19,11 @@ public class SpringWebMvcInstrumentationModule extends InstrumentationModule {
   }
 
   @Override
+  public boolean isHelperClass(String className) {
+    return className.startsWith("org.springframework.web.servlet.OpenTelemetryHandlerMappingFilter");
+  }
+
+  @Override
   public List<TypeInstrumentation> typeInstrumentations() {
     return asList(
         new WebApplicationContextInstrumentation(),
