+++ instrumentation/spring/spring-webmvc-3.1/javaagent/src/main/java/org/springframework/web/servlet/OpenTelemetryHandlerMappingFilter.java
@@ -3,12 +3,13 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package io.opentelemetry.javaagent.instrumentation.springwebmvc;
+package org.springframework.web.servlet;
 
 import static io.opentelemetry.instrumentation.api.servlet.ServerSpanNaming.Source.CONTROLLER;
 
 import io.opentelemetry.context.Context;
 import io.opentelemetry.instrumentation.api.servlet.ServerSpanNaming;
+import io.opentelemetry.javaagent.instrumentation.springwebmvc.SpringWebMvcServerSpanNaming;
 import java.io.IOException;
 import java.util.ArrayList;
 import java.util.List;
@@ -22,11 +23,9 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import org.springframework.beans.factory.support.GenericBeanDefinition;
 import org.springframework.core.Ordered;
-import org.springframework.web.servlet.HandlerExecutionChain;
-import org.springframework.web.servlet.HandlerMapping;
 import org.springframework.web.servlet.mvc.method.annotation.RequestMappingHandlerMapping;
 
-public class HandlerMappingResourceNameFilter implements Filter, Ordered {
+public class OpenTelemetryHandlerMappingFilter implements Filter, Ordered {
   private volatile List<HandlerMapping> handlerMappings;
 
   @Override
@@ -126,8 +125,8 @@ public class HandlerMappingResourceNameFilter implements Filter, Ordered {
   public static class BeanDefinition extends GenericBeanDefinition {
     public BeanDefinition() {
       setScope(SCOPE_SINGLETON);
-      setBeanClass(HandlerMappingResourceNameFilter.class);
-      setBeanClassName(HandlerMappingResourceNameFilter.class.getName());
+      setBeanClass(OpenTelemetryHandlerMappingFilter.class);
+      setBeanClassName(OpenTelemetryHandlerMappingFilter.class.getName());
     }
   }
 }
